requirements.txt
watchdog
pyspellchecker
spacy~=3.8.5
striprtf~=0.0.29
matplotlib~=3.10.1
CairoSVG~=2.7.1
beautifulsoup4~=4.13.3
pillow~=11.1.0
numpy~=2.2.4
nltk~=3.9.1
requests~=2.32.3